+++ R/waicAbund.R
@@ -19,11 +19,12 @@ waicAbund <- function(object, N.max, by.species = FALSE, ...) {
 			     'msAbund', 'lfMsAbund', 
 			     'sfMsAbund', 'msNMix', 
 			     'lfMsNMix', 'sfMsNMix', 'DS', 'spDS', 
-			     'msDS', 'lfMsDS', 'sfMsDS', 'svcAbund', 'svcMsAbund'))) {
-    stop("error: object must be one of the following classes: abund, spAbund, NMix, spNMix, msAbund, lfMsAbund, sfMsAbund, msNMix, lfMsNMix, sfMsNMix, DS, spDS, msDS, lfMsDS, sfMsDS, svcAbund, svcMsAbund\n")
+			     'msDS', 'lfMsDS', 'sfMsDS', 'svcAbund', 'svcMsAbund', 
+           'dynAbund', 'spDynAbund'))) {
+    stop("error: object must be one of the following classes: abund, spAbund, NMix, spNMix, msAbund, lfMsAbund, sfMsAbund, msNMix, lfMsNMix, sfMsNMix, DS, spDS, msDS, lfMsDS, sfMsDS, svcAbund, svcMsAbund, dynAbund, spDynAbund\n")
   }
 
-  if (!(class(object) %in% c('abund', 'spAbund', 'msAbund', 'lfMsAbund', 'sfMsAbund', 'svcAbund', 'svcMsAbund'))) {
+  if (!(class(object) %in% c('abund', 'spAbund', 'msAbund', 'lfMsAbund', 'sfMsAbund', 'svcAbund', 'svcMsAbund', 'dynAbund', 'spDynAbund'))) {
     if (missing(N.max)) {
       message("N.max not specified. Setting upper index of integration of N to 10 plus\nthe largest estimated abundance value at each site in object$N.samples")
       if (class(object) %in% c('msNMix', 'lfMsNMix', 'sfMsNMix', 
@@ -110,6 +111,17 @@ waicAbund <- function(object, N.max, by.species = FALSE, ...) {
       out <- c(elpd, pD, -2 * (elpd - pD))
       names(out) <- c("elpd", "pD", "WAIC")
     }
+  }
+  
+  # Single-species dynamic GLMs -------------------------------------------
+  if (class(object) %in% c('dynAbund', 'spDynAbund')) {
+    if (object$dist == 'zi-Gaussian') {
+      message("Calculated WAIC is only for stage 2 of the hurdle model\n")
+    }
+    elpd <- sum(apply(object$like.samples, c(2, 3), function(a) log(mean(a))), na.rm = TRUE)
+    pD <- sum(apply(object$like.samples, c(2, 3), function(a) var(log(a))), na.rm = TRUE)
+    out <- c(elpd, pD, -2 * (elpd - pD))
+    names(out) <- c("elpd", "pD", "WAIC")
   }
 
   # Single-species N-mixture models ---------------------------------------